Refactoring/create-switchswitch-common (#9)
