--- api-gateway/src/test/java/com/rljj/apigateway/ApiGatewayApplicationTests.java
@@ -1,13 +1,0 @@
-package com.rljj.apigateway;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ApiGatewayApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
