+++ crawling-service/src/main/java/com/rljj/switchswitchcrawling/SwitchswitchcrawlingApplication.java
@@ -4,11 +4,11 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.scheduling.annotation.EnableScheduling;
 
-@SpringBootApplication
-@EnableScheduling
+//@SpringBootApplication
+//@EnableScheduling
 public class SwitchswitchcrawlingApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SwitchswitchcrawlingApplication.class, args);
-	}
+//
+//	public static void main(String[] args) {
+//		SpringApplication.run(SwitchswitchcrawlingApplication.class, args);
+//	}
 }
