+++ member-service/src/main/resources/application-db.yml
@@ -1,9 +1,9 @@
 spring:
   datasource:
     driver-class-name: com.mysql.cj.jdbc.Driver
-    url: ${SPRING_DATASOURCE_URL}
-    username: root
-    password: ${SPRING_DATASOURCE_PASSWORD}
+    url: ${db.url}
+    username: ${db.username}
+    password: ${db.password}
   jpa:
     database: mysql
     database-platform: org.hibernate.dialect.MySQL8Dialect
@@ -14,4 +14,4 @@ spring:
         show_sql: true
         format_sql: true
   config:
-    import: application-secret.properties
+    import: application-secret.yml
