--- member-service/src/main/resources/application-secret.properties.txt
@@ -1,3 +1,0 @@
-JWT.SECRET=switchswitchtestkeyswitchswitchtestkeyswitchswitchtestkeyswitchswitchtestkey
-SPRING_DATASOURCE_URL=jdbc:mysql://localhost:3306/switchswitch?useSSL=false&allowPublicKeyRetrieval=true&serverTimezone=UTC&characterEncoding=UTF-8
-SPRING_DATASOURCE_PASSWORD=
