+++ member-service/src/main/resources/application-secret.yml.txt
@@ -1,0 +1,6 @@
+db:
+  url: jdbc:mysql://localhost:3306/switchswitch?serverTimezone=Asia/Seoul
+  username: root
+  password: '0000'
+jwt:
+  secret: switchswitchtestkeyswitchswitchtestkeyswitchswitchtestkeyswitchswitchtestkey
