+++ member-service/src/main/resources/application.yml
@@ -7,6 +7,7 @@ spring:
   config:
     import:
       - application-db.yml
+      - application-secret.yml
 
 eureka:
 #  instance:
@@ -28,4 +29,3 @@ management:
         include: "*"
 
 
-
