+++ settings.gradle
@@ -3,3 +3,4 @@ include 'member-service'
 include 'crawling-service'
 include 'api-gateway'
 include 'eureka-server'
+include 'switchswitch-common'
