+++ switchswitch-common/build.gradle
@@ -1,0 +1,13 @@
+dependencies {
+	implementation 'org.springframework.boot:spring-boot-starter-web'
+	implementation 'io.jsonwebtoken:jjwt-api:0.12.3'
+	implementation 'org.springframework.data:spring-data-commons'
+	implementation 'jakarta.persistence:jakarta.persistence-api'
+	implementation 'org.springframework.data:spring-data-jpa'
+	runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.3'
+	runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.3'
+	compileOnly 'org.projectlombok:lombok'
+	annotationProcessor 'org.projectlombok:lombok'
+	testImplementation 'org.springframework.boot:spring-boot-starter-test'
+	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
+}
