+++ switchswitch-common/src/main/java/com/rljj/switchswitchcommon/baseentity/BaseEntity.java
@@ -1,4 +1,4 @@
-package com.rljj.switchswitchmemberservice.global.baseentity;
+package com.rljj.switchswitchcommon.baseentity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
@@ -19,7 +19,8 @@ import java.time.LocalDateTime;
 @MappedSuperclass
 public abstract class BaseEntity {
 
-    @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
     @CreatedDate
