+++ switchswitch-common/src/main/java/com/rljj/switchswitchcommon/exception/DuplicatedException.java
@@ -1,4 +1,4 @@
-package com.rljj.switchswitchmemberservice.global.config.exception;
+package com.rljj.switchswitchcommon.exception;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
